+++ packages/eslint-config/rules/flow.js
@@ -1,5 +1,0 @@
-module.exports = {
-  extends: ['plugin:flowtype/recommended'],
-  plugins: ['flowtype'],
-  rules: {}
-};